Add in mocks for result parsers for LoginService Add in new code to update the LoginService based on Michal's recommendations.
For test login service you should test result in handler function and also have test that check
if parser was used for parsing (as you have that method in the API).
So additional Mock Parser registering result parsing.
Same would go for Publish service.
BTW. in BaseMockRestClient you have lastUploadRequest which should be renamed
but that's just a little thing.